+++ src/game.C
@@ -788,7 +788,7 @@ void /*_cdecl*/ GameSession::show_message(int bot, const char *str, ...)
   va_end(ap);
 
   int poly[10];
-  int l = strlen(buf) * 6;
+  int l = (int) strlen(buf) * 6;
   int y = bot ?  SCREEN_HEIGHT - 8 : 0;
   int yd = bot ? SCREEN_HEIGHT - 1 : 7;
   if (l > 0)
