+++ src/simfile.C
@@ -46,8 +46,8 @@ FILE *simlog = stdout;
 const char *build_libpath(const char *filename)
 {
   static char tmppath[200];
-  int l = strlen(filename);
-  int l2 = strlen(lib_path);
+  int l = (int) strlen(filename);
+  int l2 = (int) strlen(lib_path);
   if ((l + l2 + 2) > 200)
     error_jump("Can't build path, length too long: %s %s",lib_path,filename);
 #ifdef SABREWIN
@@ -55,7 +55,7 @@ const char *build_libpath(const char *filename)
 #else
   sprintf(tmppath,"%s/%s",lib_path,filename);
 #endif
-  l = strlen(tmppath);
+  l = (int) strlen(tmppath);
   if (tmppath[l-1] == '\n' || tmppath[l-1]=='\r')
     tmppath[l-1] = 0;
   return(tmppath);
@@ -138,16 +138,16 @@ int get_line(std::istream &is, char *buff, int size)
 	{
 	  if (buff[0] != '*' && buff[0] != '\x0')
 	    {
-	      l = strlen(buff);
+	      l = (int) strlen(buff);
 	      if (buff[l-1] == '\n')
 		buff[l-1] = 0;
-	      l = strlen(buff);
+	      l = (int) strlen(buff);
 	      if (buff[l-1] == '\r')
 		{
 		  buff[l-1] = 0;
 		  sim_printf("warning: %s contains DOS cr/lf\n");
 		}
-	      return (strlen(buff));
+	      return (int) (strlen(buff));
 	    }
 	}
     }
